Fix ship order tests
